moved Define from canmatrix to own package, with tests.
