+++ tests/test_canmatrix.py
@@ -359,68 +359,7 @@ def test_frame_find_wrong_signal_group(empty_frame):
     assert empty_frame.signal_group_by_name("wrong") is None
 
 
-# Define tests
-def test_define_set_default():
-    define = canmatrix.canmatrix.Define("")
-    define.set_default("string")
-    assert define.defaultValue == "string"
-    define.set_default('"quoted_string"')
-    assert define.defaultValue == "quoted_string"
 
-
-def test_define_update_enum_definition():
-    define = canmatrix.canmatrix.Define("")
-    define.type = "ENUM"
-    define.values = ["ready", "off"]
-    define.update()
-    assert define.definition == 'ENUM "ready","off"'
-
-
-def test_define_update_ingored_non_enum():
-    def_str = "INT 0 100"
-    define = canmatrix.canmatrix.Define(def_str)
-    define.update()
-    assert define.definition == def_str
-
-
-def test_define_for_int():
-    define = canmatrix.canmatrix.Define("INT -5 10")
-    assert define.type == "INT"
-    assert define.min == -5
-    assert define.max == 10
-
-
-def test_define_for_hex():
-    define = canmatrix.canmatrix.Define("HEX 0 255")
-    assert define.type == "HEX"
-    assert define.min == 0
-    assert define.max == 255
-
-
-def test_define_for_string():
-    define = canmatrix.canmatrix.Define("STRING")
-    assert define.type == "STRING"
-    assert define.min is None
-    assert define.max is None
-
-
-def test_define_for_enum():
-    define = canmatrix.canmatrix.Define('ENUM red, green')
-    assert define.type == "ENUM"
-    assert define.values == ["red", "green"]
-
-
-def test_define_for_enum_strip_quotes():
-    define = canmatrix.canmatrix.Define('ENUM "red", "green"')
-    assert define.type == "ENUM"
-    assert define.values == ["red", "green"]
-
-
-def test_define_for_float():
-    define = canmatrix.canmatrix.Define("FLOAT -2.2 111.11")
-    assert define.type == "FLOAT"
-    assert define.min == decimal.Decimal('-2.2')
-    assert define.max == decimal.Decimal('111.11')
